+++ CHANGELOG.md
@@ -8,6 +8,26 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 * Support for signed and encrypted UserInfo response. #305
 * Support for signed and encrypted ID Token. #305
 
+## [0.9.10]
+
+## Fixed
+
+* `private_key_jwt` and `client_secret_jwt` need to explicitly be enabled #331
+
+
+## [0.9.9]
+
+### Added
+
+* Added support for back-channel logout. #302
+* Added support for `private_key_jwt` Client Authentication method #322
+* Added support for `client_secret_jwt` Client Authentication method #324
+* Added PS512 encryption support #342
+
+## Fixed
+
+* Harden self-signed JWK header usage. #323
+
 ## [0.9.8]
 
 ## Fixed
@@ -60,13 +80,13 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 * it is now possible to disable upgrading from HTTP to HTTPS for development purposes by calling `setHttpUpgradeInsecureRequests(false)` #241
 * bugfix in getSessionKey when _SESSION key does not exist #251
 * Added scope parameter to refresh token request #225
-* bugfix in verifyJWTclaims when $accessToken is empty and $claims->at_hash is not #276
+* bugfix in `verifyJWTclaims` when $accessToken is empty and $claims->at_hash is not #276
 * bugfix with the `empty` function in PHP 5.4 #267
 
 ## [0.9.2]
 
 ### Added
-* Support for [PKCE](https://tools.ietf.org/html/rfc7636). Currently the supported methods are 'plain' and 'S256'.
+* Support for [PKCE](https://tools.ietf.org/html/rfc7636). Currently, the supported methods are 'plain' and 'S256'.
 
 ## [0.9.1]
 
@@ -129,7 +149,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 * Add option to send additional registration parameters like post_logout_redirect_uris. #140
 
 ### Changed
-* disabled autoload for Crypt_RSA + makre refreshToken() method tolerant for errors #137
+* disabled autoload for Crypt_RSA + make refreshToken() method tolerant for errors #137
 
 ### Removed
 *
@@ -139,7 +159,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Added
 * Added five minutes leeway due to clock skew between openidconnect server and client.
 * Fix save access_token from request in implicit flow authentication #129
-* verifyJWTsignature() method private -> public #126
+* `verifyJWTsignature()` method private -> public #126
 * Support for providers where provider/login URL is not the same as the issuer URL. #125
 * Support for providers that has a different login URL from the issuer URL, for instance Azure Active Directory. Here, the provider URL is on the format: https://login.windows.net/(tenant-id), while the issuer claim actually is on the format: https://sts.windows.net/(tenant-id).